+++ src/ZombieHelper.sol
@@ -4,45 +4,39 @@ pragma solidity ^0.8.24;
 import "./ZombieFeeding.sol";
 
 contract ZombieHelper is ZombieFeeding {
-    uint levelUpFee = 0.001 ether;
+    uint256 levelUpFee = 0.001 ether;
 
     function withdraw() external onlyOwner {
         address _owner = owner();
-        (bool success, ) = payable(_owner).call{value: address(this).balance}(
-            ""
-        );
+        (bool success,) = payable(_owner).call{value: address(this).balance}("");
         require(success, "Withdrawal failed");
     }
 
-    function setLevelUpFee(uint _fee) external onlyOwner {
+    function setLevelUpFee(uint256 _fee) external onlyOwner {
         levelUpFee = _fee;
     }
 
-    function levelUp(uint _zombieId) external payable {
+    function levelUp(uint256 _zombieId) external payable {
         require(msg.value == levelUpFee);
         zombies[_zombieId].level = zombies[_zombieId].level + 1;
     }
 
-    function changeName(
-        uint _zombieId,
-        string calldata _newName
-    ) external aboveLevel(2, _zombieId) onlyOwnerOf(_zombieId) {
+    function changeName(uint256 _zombieId, string calldata _newName)
+        external
+        aboveLevel(2, _zombieId)
+        onlyOwnerOf(_zombieId)
+    {
         zombies[_zombieId].name = _newName;
     }
 
-    function changeDna(
-        uint _zombieId,
-        uint _newDna
-    ) external aboveLevel(20, _zombieId) onlyOwnerOf(_zombieId) {
+    function changeDna(uint256 _zombieId, uint256 _newDna) external aboveLevel(20, _zombieId) onlyOwnerOf(_zombieId) {
         zombies[_zombieId].dna = _newDna;
     }
 
-    function getZombiesByOwner(
-        address _owner
-    ) external view returns (uint[] memory) {
-        uint[] memory result = new uint[](ownerZombieCount[_owner]);
-        uint counter = 0;
-        for (uint i = 0; i < zombies.length; i++) {
+    function getZombiesByOwner(address _owner) external view returns (uint256[] memory) {
+        uint256[] memory result = new uint256[](ownerZombieCount[_owner]);
+        uint256 counter = 0;
+        for (uint256 i = 0; i < zombies.length; i++) {
             if (ownerOf(i) == _owner) {
                 result[counter] = i;
                 counter++;
