+++ src/ZombieOwnership.sol
@@ -4,6 +4,5 @@ pragma solidity ^0.8.24;
 import "./ZombieAttack.sol";
 
 contract ZombieOwnership is ZombieAttack {
-
-  mapping (uint => address) zombieApprovals;
+    mapping(uint256 => address) zombieApprovals;
 }
